style(layout): Add some transition and duration when user changes theme chore(component): Moved the component `LoadingAnimation` from maps folder to global  style(component): Remove transition of colors in the `PharmacyToggleButtons` component  style(component): Remove transition of colors in the `DarkModeToggle` component feat(component): Rename the isFilterMobileOpen to isFilterOpen as we use this logic to the desktop too style(component): Remove transition from the header component and add some padding feat: When the user hovers the button to hide the list, we lower the opacity of the list. feat(component): Take as props the `isHoveringHideButton` and `setIsHoveringHideButton`.
Refactored the button of hiding the list with custom transitions, so to be more appealing and minimalist. Also, added a popover functionality so the user has a feedback.

chore(types): Add the types of `isHoveringHideButton` and `setIsHoveringHideButton` to the IPharmacyMapProps.

chore(docs): Add question "Τι είναι η επιλογή « Όλα τα Φαρμακεία » ;"
chore(providers): Remove the ReactQueryDevtools
chore(components): Rename the DutyStatus option, all, from "Ολα" to "Όλα τα Φαρμακεία"
chore(component): set default the first FAQ on FAQSection